requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
exceptiongroup==1.1.3
executing==2.0.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
numpy==1.26.1
openai==1.2.3
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.40
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
streamlit==1.28.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wcwidth==0.2.9
zipp==3.17.0
